requirements.txt
flask
pydantic
python-pptx
openai
pytest
